requirements.txt
numpy
scikit-learn
sentence-transformers
faiss-cpu
google-generativeai